refactor: 프로필 리팩터링 패키지 수정
