+++ src/main/java/com/example/kkm/user/profile/service/ProfileService.java
@@ -2,10 +2,9 @@ package com.example.kkm.user.profile.service;
 
 import com.example.kkm.user.exception.CustomException;
 import com.example.kkm.user.exception.ErrorCode;
-import com.example.kkm.user.profile.domain.ProfileInput;
-import com.example.kkm.user.profile.domain.model.Profile;
-import com.example.kkm.user.profile.domain.repository.ProfileRepository;
-import java.util.Optional;
+import com.example.kkm.user.profile.service.domain.ProfileInput;
+import com.example.kkm.user.profile.service.domain.model.Profile;
+import com.example.kkm.user.profile.service.domain.repository.ProfileRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
