+++ src/main/java/com/example/kkm/user/profile/service/domain/ProfileInput.java
@@ -1,4 +1,4 @@
-package com.example.kkm.user.profile.domain;
+package com.example.kkm.user.profile.service.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
